requirements.txt
Django==1.11.6
django-phonenumber-field==1.3.0
phonenumberslite==8.8.4
Pillow==4.3.0
psycopg2==2.7.3.2
python-decouple==3.1
sorl-thumbnail==12.3
django-bootstrap4
django-ckeditor==5.3.1
psycopg2==2.7.3.2
Pillow==4.3.0
python-decouple==3.1
Scrapy==1.4.0
scrapy-djangoitem==1.1.1
celery==4.1.0
redis==2.10.6
uwsgi==2.0.15